% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_leverage.R
\name{algo_leverage}
\alias{algo_leverage}
\title{algorithmic leveraging for linear regression using uniform and leverage score based subsampling of rows}
\usage{
algo_leverage(x, y, r, methods)
}
\arguments{
\item{x}{independent variable}

\item{y}{dependent variable}

\item{r}{subsample size}
}
\value{
estimation of parameter beta in y=beta*x
}
\description{
algorithmic leveraging for linear regression using uniform and leverage score based subsampling of rows
}
\examples{
e = rnorm(500,0,1)
x = rt(500, df = 6)
y = -x+e
algo_leverage(x, y, r=10, methods="UNIF")
}
