% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_coord.R
\name{elnet_coord}
\alias{elnet_coord}
\title{fits elastic net to data using coordinate descent algorithm}
\usage{
elnet_coord(x, y, lambda, alpha, tol)
}
\arguments{
\item{x}{independent variable}

\item{y}{dependent variable}

\item{lambda}{LASSO tuning}

\item{alpha}{proportion of usage in l1 penalty comparing with l2}
}
\value{
estimation of parameter beta in y=beta*x
}
\description{
fits elastic net to data using coordinate descent algorithm
}
