% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ols.R
\name{solve_ols}
\alias{solve_ols}
\title{solves a linear system using Gauss-Seidel or Jacobi(parallel) method}
\usage{
solve_ols(A, b, tol, methods, p)
}
\arguments{
\item{A}{matrix}

\item{b}{vector}

\item{tol}{tolarence}

\item{p}{number of cores to use for parallel}
}
\value{
estimation of parameter x in Ax=b
}
\description{
solves a linear system using Gauss-Seidel or Jacobi(parallel) method
}
